BF: rename _multiprocessing to _jlmultiprocessing to not overload standard joblib
otherwise strange things start to happen, e.g. joblib could start reporting
e.g. that

$> python joblib/_multiprocessing.py
joblib/_multiprocessing.py:29: UserWarning: This platform lacks a functioning sem_open implementation, therefore, the required synchronization primitives needed will not function, see issue 3770..  joblib will operate in serial mode
