+++ joblib/parallel.py
@@ -20,7 +20,7 @@ try:
 except:
     import pickle
 
-from ._multiprocessing import mp
+from ._jlmultiprocessing import mp
 if mp is not None:
     from .pool import MemmapingPool
     from multiprocessing.pool import ThreadPool
